requirements.txt
alembic==1.13.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
polars==0.20.15
pydantic==2.6.3
scikit-learn==1.4.1.post1
pre-commit==3.6.2
python-dotenv==1.0.1
prophet==1.1.5
streamlit==1.32.2
plotly==5.19.0
